revert "require secure backups"
This reverts commit 225a6e29e543f2a6c3fc17f152d384499cc71c97. Requiring
this causes Android users with backups already enabled to be required to
create an entirely new backup, which is dumb.

I'm so tired of the sub-par development of the Element Android client.
The Element Web implementation is so good, and the Element Android
implementation is so bad. This is consistently the case.